# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AutoDoctor MVP API
  version: 1.2.0-draft
  description: |
    Контракт REST API guest-first manual-first MVP AutoDoctor по ТЗ v0.8.
    Публичный browse shell и создание гостевой сессии не требуют приглашения.
    VIN необязателен и принимается только в запросах; ответы содержат nullable vin_masked.
    Реализуемый срез профиля автомобиля не вызывает внешний VIN/AI provider.
    Maintenance v3 включает активные mileage forecast, condition observations,
    plan-item UI preferences и идемпотентный batch/upsert history answers.
    Источник by-pilot-baseline-2 — внутренняя редакционная методика AutoDoctor,
    не OEM/manual; опубликованные версии остаются immutable.
    /vehicles/decode и /vehicles/confirm сохранены только как будущий контракт enrichment.
    Все ресурсы автомобиля адресуются с vehicleId; сервер проверяет текущего владельца.
    Все даты и время передаются в UTC в формате RFC 3339.
    Filament AI admin использует внутренний серверный интерфейс: публичные мобильные
    endpoints конфигурации, playground и A/B experiments в этом контракте отсутствуют.
servers:
  - url: https://api-dev.autodoctor.by/api/v1
    description: Development
  - url: https://api.autodoctor.by/api/v1
    description: Production
...
security:
  - SessionAuth: []
  - BearerAuth: []
paths:
...
  /vehicles/{vehicleId}/history:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      tags: [History]
      operationId: listServiceRecords
      summary: Получить историю обслуживания
      description: |
        Активная append-oriented хронология реальных service records, newest first.
        Это не history_answers declaration wizard.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Страница записей истории
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRecordList"
              examples:
                ru:
                  summary: Русское представление
                  value:
                    items:
                      - id: 22ceba30-40a7-4be7-b9c8-fd50d12c423c
                        vehicle_id: 8d7d82eb-619d-458d-b5fb-87c930ba4242
                        service_date: "2026-07-15"
                        mileage: {value: 84200, unit: km}
                        evidence_source: self
                        note: Масло заменено самостоятельно
                        items:
                          - {work_code: engine_oil, title: Моторное масло}
                          - {work_code: oil_filter, title: Масляный фильтр}
                        created_at: "2026-07-15T12:00:00Z"
                    meta: {page: 1, per_page: 20, total: 1, total_pages: 1}
                en:
                  summary: English presentation
                  value:
                    items:
                      - id: 22ceba30-40a7-4be7-b9c8-fd50d12c423c
                        vehicle_id: 8d7d82eb-619d-458d-b5fb-87c930ba4242
                        service_date: "2026-07-15"
                        mileage: {value: 84200, unit: km}
                        evidence_source: self
                        note: Oil changed by owner
                        items:
                          - {work_code: engine_oil, title: Engine oil}
                          - {work_code: oil_filter, title: Oil filter}
                        created_at: "2026-07-15T12:00:00Z"
                    meta: {page: 1, per_page: 20, total: 1, total_pages: 1}
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [History]
      operationId: createServiceRecord
      summary: Добавить запись обслуживания
      description: |
        Идемпотентно и атомарно добавляет service_record/items, upsert-ит latest
        history_answer=done_known для каждой работы, при наличии mileage создаёт
        MileageObservation(source=service) и публикует новый plan snapshot.
        service_date поставляет клиент (обычно сегодня); omitted mileage использует
        current confirmed mileage, если оно доступно. evidence_source по умолчанию self.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRecordCreate"
            example:
              service_date: "2026-07-20"
              work_codes: [engine_oil]
              mileage: {value: 84250, unit: km}
              evidence_source: self
              note: Самостоятельное обслуживание
      responses:
        "201":
          description: Запись создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRecordCreateResult"
              example:
                service_record:
                  id: 22ceba30-40a7-4be7-b9c8-fd50d12c423c
                  vehicle_id: 8d7d82eb-619d-458d-b5fb-87c930ba4242
                  service_date: "2026-07-20"
                  mileage: {value: 84250, unit: km}
                  evidence_source: self
                  note: Самостоятельное обслуживание
                  items:
                    - {work_code: engine_oil, title: Моторное масло}
                  created_at: "2026-07-20T09:30:00Z"
                mileage_observation:
                  id: 38ea4fc4-cf7e-4c5e-bc7c-85e97e631e0f
                  vehicle_id: 8d7d82eb-619d-458d-b5fb-87c930ba4242
                  mileage: {value: 84250, unit: km}
                  source: service
                  observed_at: "2026-07-20T09:30:00Z"
                  created_at: "2026-07-20T09:30:00Z"
                maintenance_plan_id: c3ecf7b6-eaab-4365-a9df-c1e34a822170
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Токен анонимной гостевой сессии.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Короткоживущий access-токен зарегистрированного пользователя.
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: |
        Язык локализованных текстов и ошибок: `ru` или `en`; региональные теги
        (`ru-BY`, `en-US`) нормализуются до базового языка. Для публичных запросов
        и ошибок до аутентификации отсутствие или неподдерживаемое значение даёт `ru`.
        После успешной аутентификации анонимной сессии отсутствие заголовка использует
        сохранённый `session.locale`; переданный заголовок имеет приоритет.
      schema:
        type: string
        examples: [ru, en, ru-BY, en-US]
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Уникальный UUID операции; повтор с тем же телом возвращает прежний результат.
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Error:
      description: Ошибка запроса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VEHICLE_NOT_SUPPORTED
              message: Для этого автомобиля проверенные данные пока не подготовлены.
              request_id: req_01J2Y6D8QK3R7T5V9W1X
              details:
                support_request_id: 7ea2c18f-a6cf-4c19-95c4-c4cb56a701b7
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_FAILED
              message: Проверьте заполненные поля.
              request_id: req_01J2Y6D8QK3R7T5V9W1X
              fields:
                mileage.value:
                  - Значение должно быть неотрицательным.
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              type: string
              pattern: "^[A-Z][A-Z0-9_]+$"
            message:
              type: string
            request_id:
              type: string
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            details:
              type: object
              additionalProperties: true
...
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, per_page, total, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
    MileageUnit:
      type: string
      enum: [km, mi]
    Mileage:
      type: object
      additionalProperties: false
      required: [value, unit]
      properties:
        value:
          type: integer
          minimum: 0
        unit:
          $ref: "#/components/schemas/MileageUnit"
...
    MileageObservationSource:
      type: string
      enum: [manual, service, expense, fuel]
    MileageObservation:
      type: object
      additionalProperties: false
      required: [id, vehicle_id, mileage, source, observed_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        vehicle_id:
          type: string
          format: uuid
        mileage:
          $ref: "#/components/schemas/Mileage"
        source:
          $ref: "#/components/schemas/MileageObservationSource"
        observed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    ServiceWorkCode:
      type: string
      enum: [engine_oil, oil_filter, air_filter, cabin_filter, brake_fluid, coolant, brake_system_inspection, tire_condition_inspection, coolant_inspection, spark_plugs, timing_drive, transmission_oil, brake_pads, brake_discs, other]
    ServiceEvidenceSource:
      type: string
      description: Стабильное machine enum; отображаемая подпись локализуется клиентом.
      enum: [self, workshop, previous_owner, document, unknown]
    ServiceRecordCreate:
      type: object
      additionalProperties: false
      required: [service_date, work_codes]
      properties:
        service_date:
          type: string
          format: date
          description: Клиент передаёт выбранную дату; UI default — локальное today.
        mileage:
          oneOf:
            - $ref: "#/components/schemas/Mileage"
            - type: "null"
          description: При отсутствии сервер использует current confirmed mileage, если оно есть.
        work_codes:
          type: array
          minItems: 1
          maxItems: 1
          uniqueItems: true
          description: |
            Active ServiceRecordScreen сохраняет ровно одну работу. Из event/consumable
            она предзаполнена; из global + пользователь выбирает её явно.
          items:
            $ref: "#/components/schemas/ServiceWorkCode"
        note:
          type: string
          maxLength: 4000
        evidence_source:
          type: string
          const: self
          default: self
          description: Active create фиксирует self; стабильный общий enum сохранён в response/import contracts.
...
    ServiceRecordItem:
      type: object
      additionalProperties: false
      required: [work_code, title]
      properties:
        work_code:
          $ref: "#/components/schemas/ServiceWorkCode"
        title:
          type: string
          description: RU/EN по Accept-Language; identity задаёт work_code.
    ServiceRecord:
      type: object
      additionalProperties: false
      required: [id, vehicle_id, service_date, mileage, evidence_source, note, items, created_at]
      properties:
        id:
          type: string
          format: uuid
        vehicle_id:
          type: string
          format: uuid
        service_date:
          type: string
          format: date
        mileage:
          oneOf:
            - $ref: "#/components/schemas/Mileage"
            - type: "null"
        evidence_source:
          $ref: "#/components/schemas/ServiceEvidenceSource"
        note:
          type: [string, "null"]
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ServiceRecordItem"
        created_at:
          type: string
          format: date-time
    ServiceRecordCreateResult:
      type: object
      additionalProperties: false
      required: [service_record, mileage_observation, maintenance_plan_id]
      properties:
        service_record:
          $ref: "#/components/schemas/ServiceRecord"
        mileage_observation:
          oneOf:
            - $ref: "#/components/schemas/MileageObservation"
            - type: "null"
          description: Создаётся с source=service, если resolved mileage присутствует.
        maintenance_plan_id:
          type: string
          format: uuid
    ServiceRecordList:
      type: object
      additionalProperties: false
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRecord"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...